requirements.txt
airtest==1.0.27
pocoui==1.0.76